Automatically start wifi hotspot
WIP. Works, but needs a setting to enable/disable it.
Changed the SWI application to a more general car settings application.
Will also add capability of automatically enabling BT PAN upon bluetooth
connection.

With this configuration, the device will automatically obtain data
connectivity when connected to a phone, as will anything connected to
the wifi hotspot.

Change-Id: Iec5550c69268b4418a948a850b3f830e95f51412