Reset ffmpeg when the system clock jumps far
There are two audio-related "issues" that ffmpeg has to deal with;
1) The audio drifts somewhat.
2) The system time jumps dramatically when the system clock is updated.

With parameters, we can handle one or the other, but not both at the same time, since handling problem (1) depends on NOT handling problem(2).

So to handle this, we add another thread to the camd process, and monitor for
large time jumps. If we detect a large time jump, we kill the ffmpeg process.
Now we just need to handle problem (1) with parameters.

Unfortunately, the solution to problem (1) involves activating the ffmpeg
resampler, which causes a significant CPU hit. The audio handling alone
now consumes ~10% of one core. This is far more than this process *should*
consume, however we are dealing with an ffmpeg limitation in that it is not
apparently capable of adding or removing samples to compensate for clock
drift without activating its resampler.

I am looking into adding a drift compensation routine to ffmpeg that does
not depend on their resampler.

Change-Id: Icd2944be0133747ffa1e4059c0ffd5fb107d883f